all working and cleaning up some pages :)
