+++ lib/core/custom_routes.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import 'path/firebase_dynamic_links.dart';
+import 'firebase/firebase_dynamic_links.dart';
 
 var customRoutes = <String, WidgetBuilder>{
   '/error': (context) => ErrorPage(),
