+++ lib/core/presentation/widgets/page_navigator.dart
@@ -1,21 +1,11 @@
-import 'package:firebase_dynamic_links/firebase_dynamic_links.dart';
-import 'package:firebase_messaging/firebase_messaging.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_app_badger/flutter_app_badger.dart';
-import 'package:flutter_local_notifications/flutter_local_notifications.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:flutter_svg/svg.dart';
-import 'package:rotary_nl_rye/core/bloc/repository.dart';
 import 'package:rotary_nl_rye/core/data/check_update.dart';
-import 'package:rotary_nl_rye/core/path/firebase_cloud_messaging.dart';
-import 'package:rotary_nl_rye/features/news/models/firestore_url.dart';
-import 'package:rotary_nl_rye/features/news/models/news.dart';
-import 'package:rotary_nl_rye/features/news/presentation/pages/non_pdf_news.dart';
-import 'package:rotary_nl_rye/features/news/presentation/widgets/pdf_viewer.dart';
-import 'package:rotary_nl_rye/features/settings/presentation/pages/social.dart';
-import 'package:rotary_nl_rye/main.dart';
+import 'package:rotary_nl_rye/core/firebase/firebase_cloud_messaging.dart';
+import 'package:rotary_nl_rye/core/firebase/firebase_dynamic_links.dart';
 
 import '../../../features/about/presentation/pages/about_page.dart';
 import '../../../features/contact/presentation/pages/contact_page.dart';
@@ -31,18 +21,18 @@ class PageNavigator extends StatefulWidget {
 }
 
 class _PageNavigatorState extends State<PageNavigator> {
-  late FireStoreUrl _news;
   String _appBadgeSupported = 'Unknown';
-  Repository _repo = Repository();
 
   @override
   initState() {
-    this._initDynamicLinks();
+    initDynamicLinks(context);
     appBadgeSupportedState();
+    //FCM start
     getToken();
     getInitialMessages(context);
     onMessage(context);
     onMessageOpenedApp(context);
+    //FCm end
     _removeBadge();
     super.initState();
     // versionCheck firebase remote config
@@ -153,100 +143,6 @@ class _PageNavigatorState extends State<PageNavigator> {
       ),
     );
   }
-
-// Dynamic links start
-
-  Future<void> _initDynamicLinks() async {
-    // TODO "onLink" is when the app is open and on the homescreen.
-    FirebaseDynamicLinks.instance.onLink(
-        onSuccess: (PendingDynamicLinkData? dynamicLink) async {
-      final Uri? deepLink = dynamicLink?.link;
-
-      if (deepLink != null) {
-        final uri = deepLink.path;
-        print("deepLink.path: $uri ");
-        if (deepLink.path == "/helloworld") {
-          // final id = deepLink.queryParameters["id"];
-          // final name = deepLink.queryParameters["name"];
-          Navigator.of(context)
-              .push(MaterialPageRoute(builder: (context) => SocialPage()));
-        }
-        if (deepLink.path == "/stories") {
-          // final id = deepLink.queryParameters["id"];
-          // final name = deepLink.queryParameters["name"];
-          Navigator.of(context)
-              .push(MaterialPageRoute(builder: (context) => SocialPage()));
-        }
-        if (deepLink.path == "/news") {
-          String? id = deepLink.queryParameters["id"];
-          print('news id $id');
-          List<News> _newsList = await _repo.fetchNews();
-          print('news fetched ${_newsList[int.parse(id!)].toString()}');
-          _newsList[int.parse(id)].isPdf
-              ? Navigator.push(
-                  context,
-                  MaterialPageRoute(
-                      builder: (context) => PDFPage(
-                          pdfId: _newsList[int.parse(id)],
-                          pdfUrl: _newsList[int.parse(id)].pdf!)),
-                )
-              : Navigator.push(
-                  context,
-                  MaterialPageRoute(
-                      builder: (context) =>
-                          NonPDFPage(data: _newsList[int.parse(id)])));
-        }
-        // still need some work
-        Navigator.pushNamed(context, '${deepLink.path}');
-      }
-    }, onError: (OnLinkErrorException e) async {
-      Navigator.pushNamed(context, '/error');
-    });
-
-// TODO "getInitialLink" is when the app is closed.
-    final PendingDynamicLinkData? data =
-        await FirebaseDynamicLinks.instance.getInitialLink();
-    final Uri? deepLink = data?.link;
-
-    final uri = deepLink?.path;
-
-    print("deepLink.path: $uri ");
-    if (deepLink?.path == "/helloworld") {
-      // final id = deepLink?.queryParameters["id"];
-      // final name = deepLink?.queryParameters["name"];
-      Navigator.of(context)
-          .push(MaterialPageRoute(builder: (context) => SocialPage()));
-    }
-    if (deepLink?.path == "/stories") {
-      // final id = deepLink?.queryParameters["id"];
-      // final name = deepLink?.queryParameters["name"];
-      Navigator.of(context)
-          .push(MaterialPageRoute(builder: (context) => SocialPage()));
-    }
-    if (deepLink?.path == "/news") {
-      String? id = deepLink?.queryParameters["id"];
-      List<News> _newsList = await _repo.fetchNews();
-
-      _newsList[int.parse(id!)].isPdf
-          ? Navigator.push(
-              context,
-              MaterialPageRoute(
-                  builder: (context) => PDFPage(
-                      pdfId: _newsList[int.parse(id)],
-                      pdfUrl: _newsList[int.parse(id)].pdf!)),
-            )
-          : Navigator.push(
-              context,
-              MaterialPageRoute(
-                  builder: (context) =>
-                      NonPDFPage(data: _newsList[int.parse(id)])));
-    }
-    // still need some work
-    // Navigator.pushNamed(context, '${deepLink?.path}');
-  }
-
-// Dynamic links end
-
 }
 
 void _removeBadge() {
